requirements.txt
pandas
numpy
nltk
streamlit
python-dotenv
unstructured
langchain-google-genai
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain_community
pymupdf
jq
packaging
toml
google-generativeai
